Separate and parallelise title rendering and card rendering
Previously, the `renderCards()` function would make a request for both fronter information and system information, and then update both at the same time. This is unnecessary though, because system information is only used to update the titles, and fronter information is only used to update the cards.

Now, they're separated into their own functions and run at the same time. This means that they'll update as soon as they're able to instead of waiting for the other one to finish, meaning the time to first update is slightly lower.

In practice this is an improvement of like sometimes 50ms, so it's kind of pointless... But it *is* still an improvement! It also makes the code a bit easier to understand by separating things.

As an extension to that last thing, the `backButton()` call got moved to outside of the `renderCards()` function because it didn't really make sense there.