Handle errors better
Instead of returning `null`s and passing them all over the place and doing lots of if checks, instead we do it the proper way with throw/catch. This means *all* errors are caught now, not just some of the ones we predicted.

`showInput()` was changed to use a switch statement instead of lots of `if` and `else if`s, which is easier to read (imo). If the error is a predicted one it'll show a nice custom error message, but if it's an unpredicted one it'll tell you it's an unknown error and tell you what the error actually is.