Make the error input use it's own dedicated container
Previously, it would replace the container contents with the error. This used to be fine, but now rendering titles and cards happen independently and one can error while the other doesn't. This means that the cards could be render successfully, and then get replaced by an error when the title updating fails.

Overwriting the cards container is how the "Loading fronters..." text got removed though. So now that's done with a `.catch()` on the `renderCards()` promise, which also re-throws the error so it can be handled like it would if the `updateTitles()` errored.

Top-level await also doesn't seem to work in top-level code blocks, so the main code is wrapped in an async function that immediately executes itself.