Add support for loading page when there's no switches logged
PluralKit's API will return `HTTP 204 (No Content)` if there's no switches logged and you request a system's fronters, instead of just returning an empty array like it does when there's no fronters but at least one switch logged.

I don't know why it does this, but it does. It's not in the documentation. Maybe a bug?